pyproject.toml
[project]
name = "email-investigation-tool"
version = "0.1.0"
description = "SMTP debugging and ISP interference detection tool for large email attachments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "python-gnupg>=0.5.0",
    "dnspython>=2.4.0",
    "scapy>=2.5.0",
    "psutil>=5.9.0",
]
